fix(distributed): int-coerce remote candidate bounds; backport cascor-core normalization (#319)
Root cause (dual-path remote candidate crash): _dispatch_to_remote_workers
serialized the int-valued candidate bounds random_max_value / sequence_max_value
with float(), so the JSON wire carried 1.0 instead of 1. The remote worker then
fed those floats to random.randint() / range() and raised "'float' object cannot
be interpreted as an integer", starving the remote leg into the local-retry
fallback. Send them as int().

Backport (juniper-cascor-core drift invariant): the worker-payload normalization
that juniper-ml#370-376 added directly to the juniper-cascor-core PACKAGE
(CandidateUnit._coerce_int_like + ActivationWithDerivative._normalize_activation_fn)
is now backported verbatim into cascor src, restoring the package==src invariant
the drift-guard (juniper-ml tests/test_cascor_core_drift.py) enforces. cascor src
is the single source of truth the package extracts from, so the behavior now lives
where it belongs and is tested at the source. The dispatch int() fix makes the
model-side coercion belt-and-suspenders rather than load-bearing.

Sync point: candidate_unit.py / activation.py are byte-identical to the package at
juniper-ml@f4d66db (#384).

Changes:
- src/cascade_correlation/cascade_correlation.py: float()->int() for the two remote
  candidate bounds in the dispatch payload.
- src/candidate_unit/candidate_unit.py, src/utils/activation.py: backport the
  worker-facing normalization (byte-identical to the juniper-cascor-core package).
- src/tests/unit/test_candidate_result_collection_dualpath_regression.py: guard that
  dispatched bounds stay int on the wire.
- src/tests/unit/test_candidate_core_worker_contracts.py (new): source-side contracts
  for the worker-facing normalization, mirroring the package smoke tests.

Testing:
- Full cascor unit suite: 3689 passed, 0 failed, 15 skipped (--slow).
- Serialization/pickle (--integration --slow): 62 passed (backport touches
  __getstate__/__setstate__; cascor's local pool pickles candidates).
- New contracts + dual-path regression: 18 passed.
- juniper-ml drift-guard re-run against this branch: passes (invariant restored).
- pre-commit (black/isort/flake8/mypy/bandit): clean.

+++ src/cascade_correlation/cascade_correlation.py
@@ -1217,8 +1217,12 @@ class CascadeCorrelationNetwork:
                         "random_value_scale": float(random_value_scale),
                         "candidate_uuid": candidate_uuid,
                         "candidate_seed": candidate_seed,
-                        "random_max_value": float(random_max_value),
-                        "sequence_max_value": float(sequence_max_value),
+                        # Remote candidate bounds are int-valued and feed random.randint()/
+                        # range() in CandidateUnit (which reject floats). Send them as int so
+                        # the JSON wire preserves integer-ness — float() here made the remote
+                        # worker raise "'float' object cannot be interpreted as an integer".
+                        "random_max_value": int(random_max_value),
+                        "sequence_max_value": int(sequence_max_value),
                     },
                     "training_params": {
                         "epochs": int(self.candidate_epochs),
